Fix baseOffset calculations in PDFWriter.computerBufferSize
Before, `baseOffset` was being used as the initial value of `size`, so
that the xref calculations would come out at the right offsets.  As a
side effect, the final size was always `+baseOffset` too large, which
resulted in over-allocation in `serializeToBuffer`.

Now, `baseOffset` is only added at the times when xref calculations are
being done, and the final calculated size represents just the new data
being added to the file.